add gard and auth
