+++ routes/api/contacts/index.js
@@ -1,6 +1,6 @@
 const express = require("express");
 const router = express.Router();
-const Contacts = require("../../../controllers/contacts");
+const ctrl = require("../../../controllers/contacts");
 const guard = require("../../../helpers/guard");
 
 const {
@@ -9,84 +9,21 @@ const {
   validationUpdateStatusContact,
 } = require("./validation");
 
-router.get("/", async (req, res, next) => {
-  try {
-    const contacts = await Contacts.listContacts();
-    return res.json({ status: "success", code: 200, data: { contacts } });
-  } catch (error) {
-    next(e);
-  }
-});
+router.get("/", guard, ctrl.getAllContacts);
 
-router.get("/:id", async (req, res, next) => {
-  try {
-    const contact = await Contacts.getContactById(req.params.id);
-    if (contact) {
-      return res.json({ status: "success", code: 200, data: { contact } });
-    }
-    return res.json({ status: "error", code: 404, message: "Not Found!" });
-  } catch (e) {
-    next(e);
-  }
-});
+router.get("/:id", guard, ctrl.getContactById);
 
-router.post("/", validationAddContact, async (req, res, next) => {
-  try {
-    const contact = await Contacts.addContact(req.body);
-    return res
-      .status(201)
-      .json({ status: "success", code: 201, data: { contact } });
-  } catch (e) {
-    next(e);
-  }
-});
+router.post("/", guard, validationAddContact, ctrl.addContact);
 
-router.delete("/:id", async (req, res, next) => {
-  try {
-    const contact = await Contacts.removeContact(req.params.id);
-    if (contact) {
-      return res.json({
-        status: "success",
-        code: 200,
-        data: { contact },
-        message: "contact deleted",
-      });
-    }
-    return res.json({ status: "error", code: 404, message: "Not found" });
-  } catch (e) {
-    next(e);
-  }
-});
+router.delete("/:id", guard, ctrl.deleteById);
 
-router.put("/:id", validationUpdateContact, async (req, res, next) => {
-  try {
-    const contact = await Contacts.updateContact(req.params.id, req.body);
-    if (contact) {
-      return res.json({ status: "success", code: 200, data: { contact } });
-    }
-    return res.json({ status: "error", code: 404, message: "Not found" });
-  } catch (e) {
-    next(e);
-  }
-});
+router.put("/:id", guard, validationUpdateContact, ctrl.updateContact);
 
 router.patch(
   "/:id/favorite",
   validationUpdateStatusContact,
-  async (req, res, next) => {
-    try {
-      const contact = await Contacts.updateStatusContact(
-        req.params.id,
-        req.body
-      );
-      if (contact) {
-        return res.json({ status: "success", code: 200, data: { contact } });
-      }
-      return res.json({ status: "error", code: 404, message: "Not found" });
-    } catch (e) {
-      next(e);
-    }
-  }
+  guard,
+  ctrl.updateStatusContact
 );
 
 module.exports = router;
